+++ src/utils/nFormatter.ts
@@ -1,0 +1,12 @@
+export function nFormatter(num:number) { //조회수 'k m'
+  if (num >= 1000000000) {
+     return (num / 1000000000).toFixed(1).replace(/\.0$/, '') + 'G';
+  }
+  if (num >= 1000000) {
+     return (num / 1000000).toFixed(1).replace(/\.0$/, '') + 'M';
+  }
+  if (num >= 1000) {
+     return (num / 1000).toFixed(1).replace(/\.0$/, '') + 'K';
+  }
+  return num;
+}
